Remove video metadata as option and make run all the time.
