+++ cdx_writer/dispatcher.py
@@ -1,5 +1,5 @@
 from .handler import (RecordHandler, ResponseHandler, RevisitHandler,
-                      ResourceHandler, FtpHandler, WarcinfoHandler)
+                      ResourceHandler, FtpHandler, WarcinfoHandler, VideoMetaHandler)
 
 __all__ = [
     'RecordDispatcher', 'DefaultDispatcher', 'AllDispatcher'
@@ -71,6 +71,13 @@ class DefaultDispatcher(RecordDispatcher):
         elif record.url.startswith(('http://', 'https://')):
             return ResourceHandler
         return None
+    
+    def dispatch_metadata(self, record, env):
+        content_type = record.content_type
+
+        if content_type and content_type.startswith('application/json;generator-youtube-dl'):
+            return VideoMetaHandler
+        return None
 
 class AllDispatcher(DefaultDispatcher):
 
@@ -86,6 +93,13 @@ class AllDispatcher(DefaultDispatcher):
 
     def dispatch_warcinfo(self, record, env):
         return WarcinfoHandler
+
+    def dispatch_metadata(self, record, env):
+        content_type = record.content_type
+
+        if content_type and content_type.startswith('application/json;generator-youtube-dl'):
+            return VideoMetaHandler
+        return None
 
     def dispatch_any(self, record, env):
         return RecordHandler
